Refactor Credential Management
