+++ src/fidokey/credential_management/credential_management_command.rs
