+++ src/fidokey/make_credential/make_credential_params.rs
@@ -2,13 +2,13 @@
 make_credential API parameters
 */
 
-use crate::credential_management_params::CredentialProtectionPolicy;
 use crate::public_key::PublicKey;
 use crate::public_key_credential_descriptor::PublicKeyCredentialDescriptor;
 use crate::public_key_credential_user_entity::PublicKeyCredentialUserEntity;
 use crate::str_buf::StrBuf;
 
 use super::make_credential_params::Extension as Mext;
+use super::CredentialProtectionPolicy;
 
 use std::fmt;
 use strum_macros::AsRefStr;
